fixup! move loader to src
