+++ packages/babel-register/test/fixtures/experimental-loader/package.json
@@ -1,3 +1,1 @@
-{
-  "type": "module"
-}
+{ "type": "module" }
